+++ bin/nanocl/src/commands/state.rs
@@ -221,7 +221,7 @@ async fn attach_to_cargoes(
   Ok(())
 }
 
-fn gen_client(meta: &StateMeta) -> IoResult<NanocldClient> {
+fn gen_client(host: &str, meta: &StateMeta) -> IoResult<NanocldClient> {
   let client = match meta.api_version.clone() {
     api_version if meta.api_version.starts_with("http") => {
       let mut paths = api_version
@@ -238,7 +238,7 @@ fn gen_client(meta: &StateMeta) -> IoResult<NanocldClient> {
       NanocldClient::connect_with_url(&url, version)
     }
     api_version if meta.api_version.starts_with('v') => {
-      NanocldClient::connect_with_unix_version(&api_version)
+      NanocldClient::connect_with_url(host, &api_version)
     }
     _ => {
       let mut paths = meta
@@ -380,10 +380,10 @@ async fn parse_state_file(
   read_from_file(&path)
 }
 
-async fn exec_state_apply(opts: &StateOpts) -> IoResult<()> {
+async fn exec_state_apply(host: &str, opts: &StateOpts) -> IoResult<()> {
   let (meta, yaml) = parse_state_file(&opts.file_path).await?;
 
-  let client = gen_client(&meta)?;
+  let client = gen_client(host, &meta)?;
   let args = parse_build_args(&yaml, opts.args.clone())?;
   let mut namespace = String::from("default");
   let mut cargoes = Vec::new();
@@ -449,9 +449,9 @@ async fn exec_state_apply(opts: &StateOpts) -> IoResult<()> {
   Ok(())
 }
 
-async fn exec_state_revert(opts: &StateOpts) -> IoResult<()> {
+async fn exec_state_revert(host: &str, opts: &StateOpts) -> IoResult<()> {
   let (meta, yaml) = parse_state_file(&opts.file_path).await?;
-  let client = gen_client(&meta)?;
+  let client = gen_client(host, &meta)?;
   let args = parse_build_args(&yaml, opts.args.clone())?;
   let yaml = inject_data(yaml.clone(), &args, &client).await?;
   let data = serde_json::to_value(&yaml)
@@ -480,9 +480,9 @@ async fn exec_state_revert(opts: &StateOpts) -> IoResult<()> {
   Ok(())
 }
 
-pub async fn exec_state(args: &StateArgs) -> IoResult<()> {
+pub async fn exec_state(host: &str, args: &StateArgs) -> IoResult<()> {
   match &args.commands {
-    StateCommands::Apply(opts) => exec_state_apply(opts).await,
-    StateCommands::Revert(opts) => exec_state_revert(opts).await,
+    StateCommands::Apply(opts) => exec_state_apply(host, opts).await,
+    StateCommands::Revert(opts) => exec_state_revert(host, opts).await,
   }
 }